QA: QA 사항 수정

+++ DDANZI_iOS/DDANZI_iOS/Presentation/ProductDetail/ViewControllers/OptionSelect/OptionSelectViewController.swift
@@ -148,8 +148,13 @@ final class OptionSelectViewController: UIViewController {
       bottomButton.heartButton.rx.tap
           .bind(with: self) { owner, _ in
               Amplitude.instance().logEvent("click_detail_heart")
+            if !(UserDefaults.standard.bool(forKey: .isLogin)) {
+                owner.view.showToast(message: "로그인 이후 이용 가능합니다.", at: 130.adjusted)
+                owner.navigationController?.pushViewController(LoginViewController(signUpFrom: "like"), animated: false)
+            } else {
               let id = owner.productId
               owner.isInterest ? owner.deleteInterest(id: id) : owner.addInterest(id: id)
+            }
           }
           .disposed(by: disposeBag)
   }
